+++ webcomponents/charts/CHANGELOG.md
@@ -1,5 +1,5 @@
-<a name="1.0.0-rc.2-1"></a>
-# 1.0.0-rc.2-1 (2019-10-XX)
+<a name="1.0.0-rc.3"></a>
+# 1.0.0-rc.2-1 (2019-10-19)
 
 ### Breaking changes
 
@@ -9,6 +9,7 @@
 ### Features
 
 * add a new CSS4 variable `--deckgo-chart-stroke` in case you would like to use a default styling color for all the graphs' lines
+* update most recent libs
 
 <a name="1.0.0-rc.2"></a>
 # 1.0.0-rc.2 (2019-10-09)